+++ R/fcat.R
@@ -160,6 +160,7 @@ as.fcat.data.frame <- function(x) {
       ret[[nm]] <- eval(str2lang(df[1, "Expression"]))
     }
   }
+
   
   return(ret)
   
@@ -441,15 +442,54 @@ print.fcat <- function(x, ..., verbose = FALSE) {
    cat(grey60("# A format catalog: " %+% 
                 as.character(length(x)) %+% " formats\n"))
     
+     
+   #dat <- as.data.frame(x)
+   for(nm in names(x)) {
+     
+     ob <- x[[nm]]
+
+     if (any(class(ob) == "fmt")) {
+       cat(paste0("- $", nm, ": type U, ", length(ob),  " conditions\n"))
+       
+       
+     }  else if (any(class(x[[nm]]) == "function")) {
+        
+       cat(paste0("- $", nm, ": type F, ",  "1 function \n"))
+       
+     } else if (any(class(x[[nm]]) %in% c("character", "numeric", 
+                                          "integer", "Date", 
+                                          "POSIXct", "POSIXlt"))) {
+       
+       if (length(ob) > 1)
+         cat(paste0("- $", nm, ": type V, ", length(ob),  " elements\n"))
+       else 
+         cat(paste0("- $", nm, ": type S, \"", ob,"\"\n"))
+     }
+   }
     
-   dat <- as.data.frame(x)
-   
-   print(dat)
    
   }
     
   invisible(x)
 }
+
+
+# grey60 <- make_style(grey60 = "#999999")
+# cat(grey60("# A format catalog: " %+% 
+#              as.character(length(x)) %+% " formats\n"))
+# 
+# 
+# dat <- as.data.frame(x)
+# if (!is.null(row_limit)) {
+#   if (nrow(dat) > row_limit) {
+#     dat1 <- dat[1:row_limit, ]
+#     print(dat1)
+#     cat(grey60(paste("# ... with", nrow(dat) - row_limit, "more rows\n")))
+#     
+#   } else
+#     print(dat)
+# } else 
+#   print(dat)
 
 
 #' @title Class test for a format catalog